Enables Longer Timeout for Model Training
With the rise of AI it is becoming more common to do expensive computations in a strategy which might take a while. LEAN currently has a limit of 10 minutes per loop of the algorithm designed to prevent an algorithm locking up and crashing without user awareness.

`QCAlgorithm.TrainingCompleted` is an wait event handle that signals that the a work is being performed by the `QCAlgorithm.Train`. The `AlgorithmManager` is aware of this special algorithm state and use a larger loop timeout set by `Controls`.